renaming + input params validation
