+++ README.MD
@@ -1,3 +1,3 @@
-# Delayed queue based on redis
+# Delayed queue
 
 [![Build Status](https://travis-ci.com/fred84/delayedQueue.svg?branch=master)](https://travis-ci.com/fred84/delayedQueue)
