+++ build.gradle
@@ -21,6 +21,7 @@ dependencies {
     implementation "ch.qos.logback:logback-classic:1.2.3"
     implementation "javax.annotation:javax.annotation-api:1.3.2"
     implementation "io.micrometer:micrometer-core:1.3.0"
+    implementation 'org.jetbrains:annotations:16.0.1'
 
     testImplementation "org.junit.jupiter:junit-jupiter:5.5.2"
     testImplementation "org.hamcrest:hamcrest:2.1"
